# swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Engineering-grade backend for property management.",
        "title": "ShiftProof API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@shiftproof.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/contact": {
            "post": {
                "description": "Accepts a contact form submission and logs it. Returns 200 immediately.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Submit contact form",
                "operationId": "submitContact",
                "parameters": [
                    {
                        "description": "Contact form",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ContactRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                },
                "x-google-backend": {
                    "address": "${CLOUD_RUN_URL}/api/v1/contact",
                    "path_translation": "CONSTANT_ADDRESS"
                }
            }
        },
...
        "/api/v1/users/onboarding": {
            "post": {
                "security": [
                    {
                        "firebase": []
                    }
                ],
                "description": "Sets the user's profile on first-time onboarding. name and gender are required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Complete onboarding",
                "operationId": "onboarding",
                "parameters": [
                    {
                        "description": "Onboarding Data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OnboardingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AppUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                },
                "x-google-backend": {
                    "address": "${CLOUD_RUN_URL}/api/v1/users/onboarding",
                    "path_translation": "CONSTANT_ADDRESS"
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AppUser": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "string"
                },
                "authIdentifier": {
                    "description": "Firebase UID",
                    "type": "string"
                },
                "avatarUrl": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "createdAt": {
                    "description": "\"YYYY-MM-DD\"",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "description": "MALE, FEMALE, CO_LIVING",
                    "type": "string"
                },
                "gstin": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pan": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profileCompleted": {
                    "type": "boolean"
                },
                "providers": {
                    "description": "[\"google\", \"phone\"] — active providers only",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "roles": {
                    "description": "\"OWNER\", \"TENANT\" — global account type",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.ContactRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.OnboardingRequest": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "firebase": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://securetoken.google.com/${FIREBASE_PROJECT_ID}",
            "scopes": {
                "openid": "access to identity"
            },
            "x-google-audiences": "${FIREBASE_PROJECT_ID}",
            "x-google-issuer": "https://securetoken.google.com/${FIREBASE_PROJECT_ID}",
            "x-google-jwks_uri": "https://www.googleapis.com/service_accounts/v1/metadata/x509/securetoken@system.gserviceaccount.com"
        }
    },
...
}